case : change page + size type int64 -> int32

+++ internal/app/close_reason.go
@@ -103,8 +103,8 @@ func (s *CloseReasonService) ListCloseReasons(ctx context.Context, req *_go.List
 		Fields:  fields,
 		Context: ctx,
 		Sort:    req.Sort,
-		Page:    int64(page),
-		Size:    int64(req.Size),
+		Page:    int32(page),
+		Size:    int32(req.Size),
 		Time:    t,
 	}
