+++ internal/app/service.go
@@ -143,8 +143,8 @@ func (s *ServiceService) ListServices(ctx context.Context, req *api.ListServiceR
 		Session: session,
 		Context: ctx,
 		Sort:    req.Sort,
-		Page:    int64(page),
-		Size:    int64(req.Size),
+		Page:    int32(page),
+		Size:    int32(req.Size),
 		Time:    t,
 		Filter:  make(map[string]interface{}),
 	}
